requirements.txt
pyserial
time